Add dirty check.

+++ make-zip-mac.sh
@@ -1,4 +1,10 @@
-#!/bin/bash -ex
+#!/bin/bash -e
+if [[ $(git status --porcelain chrome) != "" ]]; then
+  echo "Working tree is dirty:"
+  git status --porcelain chrome
+  exit 1
+fi
+set -x
 V=$(cat chrome/manifest.json | grep '"version"' | grep -o "\d*\.\d*\.\d*")
 rm -f "aws-console-fixes-$V.zip"
 cd chrome
